Improved AuthorViewController & added AuthorViewModel
