+++ FakeNFT/Screens/Catalog/View/Collection/AuthorViewModel.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+final class AuthorViewModel {
+    let authorPageURL: String
+    @Observable private(set) var loadingInProgress: Bool = false
+    
+    init(authorPageURL: String) {
+        self.authorPageURL = authorPageURL
+    }
+    
+    func changeLoadingStatus(loadingInProgress: Bool) {
+        self.loadingInProgress = loadingInProgress
+    }
+}
